ISAAC ADDED
Previously Isaac has been selectable in the menu , but its functions have been disabled. Everything is now functional

FEATURES
- INFINITE HEALTH
- INFINITE BOMBS
- MAX DAMAGE
- MAX FIRE RATE
- INFINITE COINS

Keep in mind all options are for "The Binding of Isaac: Wrath of The Lamb" only. If requested I will add support for non DLC version of the game and create a sub menu for user to select game version before patch menu is displayed